Revert "feat: Default to `NULLS LAST` (#2617)"
This reverts commit 967adb97decae5b72a3eb3e2705d17d7a532774d.
