+++ bindings/prql-python/python/tests/test_website_queries.py
@@ -1,18 +1,6 @@
 import re
 
 import prql_python as prql
-import pytest
-import yaml
-
-
-@pytest.fixture()
-def example_queries():
-    website_path = "../../web/website/content/_index.md"
-    with open(website_path, "r") as f:
-        website = f.read()
-    website_yaml = yaml.safe_load(website.replace("---", ""))
-    showcase_section = website_yaml["showcase_section"]["examples"]
-    return showcase_section
 
 
 def normalize(sql: str) -> str:
@@ -30,5 +18,5 @@ def test_all_examples(example_queries):
         compiled_normalized = normalize(compiled)
         truth_normalized = normalize(query["sql"])
         assert (
-            compiled_normalized == truth_normalized
+            truth_normalized == compiled_normalized
         ), f"Failed on Query ID: '{query['id']}'"
