+++ prql-compiler/src/sql/gen_expr.rs
@@ -719,7 +719,7 @@ pub(super) fn translate_column_sort(
         } else {
             Some(false)
         },
-        nulls_first: ctx.dialect.nulls_first(),
+        nulls_first: None,
     })
 }
